+++ packages/core/src/cli/validate.ts
@@ -1,9 +1,9 @@
 import minimist from 'minimist';
 
 import { ValidateUpgradeSafetyOptions, validateUpgradeSafety } from '.';
-import { withValidationDefaults } from '..';
 import { ValidationError, errorKinds } from '../validate/run';
 import debug from '../utils/debug';
+import { withCliDefaults } from './validate/validate-upgrade-safety';
 
 const USAGE = 'Usage: npx @openzeppelin/upgrades-core validate [<BUILD_INFO_DIR>] [<OPTIONS>]';
 const DETAILS = `
@@ -14,7 +14,8 @@ Arguments:
 
 Options:
   --contract <CONTRACT>  The name or fully qualified name of the contract to validate. If not specified, all upgradeable contracts in the build info directory will be validated.
-  --reference <REFERENCE_CONTRACT>  The name or fully qualified name of the reference contract to use for storage layout comparisons. Can only be used along with the --contract option. If not specified, uses the @custom:oz-upgrades-from annotation in the contract that is being validated.
+  --reference <REFERENCE_CONTRACT>  Can only be used when the --contract option is also provided. The name or fully qualified name of the reference contract to use for storage layout comparisons. If not specified, uses the @custom:oz-upgrades-from annotation if it is defined in the contract that is being validated.
+  --requireReference  Can only be used when the --contract option is also provided. If specified, requires either the --reference option to be provided or the contract to have a @custom:oz-upgrades-from annotation.
   --unsafeAllow "<VALIDATION_ERRORS>"  Selectively disable one or more validation errors. Comma-separated list with one or more of the following: ${errorKinds.join(
     ', ',
   )}
@@ -45,6 +46,7 @@ function parseArgs(args: string[]) {
       'unsafeSkipStorageCheck',
       'unsafeAllowCustomTypes',
       'unsafeAllowLinkedLibraries',
+      'requireReference',
     ],
     string: ['unsafeAllow', 'contract', 'reference'],
     alias: { h: 'help' },
@@ -87,10 +89,15 @@ export function getFunctionArgs(parsedArgs: minimist.ParsedArgs, extraArgs: stri
     const buildInfoDir = extraArgs.length === 1 ? undefined : extraArgs[1];
     const contract = getAndValidateString(parsedArgs, 'contract');
     const reference = getAndValidateString(parsedArgs, 'reference');
-    if (reference !== undefined && contract === undefined) {
-      throw new Error('The --reference option can only be used along with the --contract option.');
-    }
     const opts = withDefaults(parsedArgs);
+
+    if (contract === undefined) {
+      if (reference !== undefined) {
+        throw new Error('The --reference option can only be used along with the --contract option.');
+      } else if (opts.requireReference) {
+        throw new Error('The --requireReference option can only be used along with the --contract option.');
+      }
+    }
     return { buildInfoDir, contract, reference, opts };
   }
 }
@@ -117,6 +124,7 @@ function validateOptions(parsedArgs: minimist.ParsedArgs) {
         'unsafeAllow',
         'contract',
         'reference',
+        'requireReference',
       ].includes(key),
   );
   if (invalidArgs.length > 0) {
@@ -152,7 +160,8 @@ export function withDefaults(parsedArgs: minimist.ParsedArgs): Required<Validate
     unsafeAllowCustomTypes: parsedArgs['unsafeAllowCustomTypes'],
     unsafeAllowLinkedLibraries: parsedArgs['unsafeAllowLinkedLibraries'],
     unsafeAllow: getUnsafeAllowKinds(parsedArgs['unsafeAllow']),
+    requireReference: parsedArgs['requireReference'],
   };
 
-  return withValidationDefaults(allOpts);
+  return withCliDefaults(allOpts);
 }